feat(profile): implement CreationPage and ProfilePage components for profile creation flow
