+++ src/app/profile/create/[step]/page.tsx
@@ -1,10 +1,7 @@
-'use client'
-import { StepComponent } from "@/components/profile/create/StepComponent";
+import { CreationPage } from "@/components/profile/create/CreationPage";
 
 export default function page(){
   return (
-    <div className={'h-full flex mx-auto justify-center flex-col'}>
-      <StepComponent/>
-    </div>
+   <CreationPage/>
   )
 }
