+++ src/components/profile/create/StepComponent.tsx
@@ -1,12 +1,13 @@
 'use client';
-import { useParams } from "next/navigation";
-import { StartComponent } from "@/components/profile/create/StartComponent";
-import { PersonalInfoComponent } from "@/components/profile/create/PersonalInfoComponent";
-import { LocationWorkComponent } from "@/components/profile/create/LocationWorkComponent";
-import { PreferencesComponent } from "@/components/profile/create/PreferencesComponent";
-import { LifestyleComponent } from "@/components/profile/create/LifestyleComponent";
-import { ReviewComponent } from "@/components/profile/create/ReviewComponent";
+import { useParams, useRouter } from "next/navigation";
+import { StartComponent } from "@/components/profile/create/steps/StartComponent";
+import { PersonalInfoComponent } from "@/components/profile/create/steps/PersonalInfoComponent";
+import { LocationWorkComponent } from "@/components/profile/create/steps/LocationWorkComponent";
+import { PreferencesComponent } from "@/components/profile/create/steps/PreferencesComponent";
+import { LifestyleComponent } from "@/components/profile/create/steps/LifestyleComponent";
+import { ReviewComponent } from "@/components/profile/create/steps/ReviewComponent";
 import Link from "next/link";
+import { useProfileQuery } from "@/hooks/react-query/profiles";
 
 export const PROFILE_STEPS = [
   'welcome',
@@ -18,7 +19,19 @@ export const PROFILE_STEPS = [
 ];
 
 export function StepComponent() {
+  const router = useRouter()
+  const query = useProfileQuery()
   const params = useParams<{ step: string }>();
+
+  if (query.isLoading) {
+    return <div>Loading...</div>;
+  }
+
+  if (query.isSuccess && !query.data?.profile) {
+    console.log("YEP")
+    router.replace('/profile')
+  }
+
   const currentStep = params?.step || 'welcome';
 
   const profileSteps = [
