+++ src/components/profile/create/steps/LocationWorkComponent.tsx
@@ -2,7 +2,7 @@
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import { PROFILE_STEPS } from "./StepComponent";
+import { PROFILE_STEPS } from "../StepComponent";
 import { Card } from "@/components/ui/card";
 import { useState } from "react";
 import { useParams } from "next/navigation";
